+++ test/unit/lib/vehicle.js
@@ -5,7 +5,6 @@ const nock = require('nock');
 const test = require('ava');
 
 const Vehicle = require('../../../lib/vehicle');
-const BatchRequest = require('../../../lib/batch-request');
 const {USER_AGENT} = require('../../../lib/util');
 
 const VID = 'ada7207c-3c0a-4027-a47f-6215ce6f7b93';
@@ -28,8 +27,8 @@ test.afterEach(function(t) {
 });
 
 test('constructor', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .matchHeader('sc-unit-system', 'metric')
     .get('/default')
     .reply(200, 'default');
@@ -42,12 +41,11 @@ test('constructor', async function(t) {
 
   const res = await vehicle.request('/default');
   t.is(res, 'default');
-
 });
 
 test('constructor - imperial', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .matchHeader('sc-unit-system', 'imperial')
     .get('/constructor/imperial')
     .reply(200, 'imperial build');
@@ -60,7 +58,6 @@ test('constructor - imperial', async function(t) {
 
   const res = await vehicle.request('/constructor/imperial');
   t.is(res, 'imperial build');
-
 });
 
 test('constructor - errors without new', function(t) {
@@ -73,8 +70,8 @@ test('constructor - errors for invalid unit', function(t) {
 });
 
 test('setUnitSystem - imperial', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .matchHeader('sc-unit-system', 'imperial')
     .get('/unit/imperial')
     .reply(200, 'imperial');
@@ -85,12 +82,11 @@ test('setUnitSystem - imperial', async function(t) {
 
   const res = await vehicle.request('/unit/imperial');
   t.is(res, 'imperial');
-
 });
 
 test('setUnitSystem - metric', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .matchHeader('sc-unit-system', 'metric')
     .get('/unit/metric')
     .reply(200, 'metric');
@@ -101,7 +97,6 @@ test('setUnitSystem - metric', async function(t) {
 
   const res = await vehicle.request('/unit/metric');
   t.is(res, 'metric');
-
 });
 
 test('setUnitSystem - error', function(t) {
@@ -110,19 +105,18 @@ test('setUnitSystem - error', function(t) {
 });
 
 test('disconnect', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .delete('/application')
     .reply(200, 'disconnect');
 
   const response = await vehicle.disconnect();
   t.is(response, 'disconnect');
-
 });
 
 test('permissions', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/permissions')
     .reply(200, {
       permissions: ['permission1', 'permission2', 'permission3'],
@@ -130,12 +124,11 @@ test('permissions', async function(t) {
 
   const permissions = await vehicle.permissions();
   t.is(permissions.length, 3);
-
 });
 
 test('has permissions - single', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/permissions')
     .reply(200, {
       permissions: ['permission1', 'permission2', 'permission3'],
@@ -147,22 +140,24 @@ test('has permissions - single', async function(t) {
 });
 
 test('has permissions - multi', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/permissions')
     .reply(200, {
       permissions: ['permission1', 'permission2', 'permission3'],
     });
 
-  const hasPermissions = await vehicle.hasPermissions(
-    ['permission1', 'required:permission2']);
+  const hasPermissions = await vehicle.hasPermissions([
+    'permission1',
+    'required:permission2',
+  ]);
 
   t.is(hasPermissions, true);
 });
 
 test('has permissions - false', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/permissions')
     .reply(200, {
       permissions: ['permission1', 'permission2', 'permission3'],
@@ -174,38 +169,38 @@ test('has permissions - false', async function(t) {
 });
 
 test('has permissions - multi false', async function(t) {
-
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/permissions')
     .reply(200, {
       permissions: ['permission1', 'permission2', 'permission3'],
     });
 
-  const hasPermissions = await vehicle.hasPermissions(
-    ['permission1', 'permission4']);
+  const hasPermissions = await vehicle.hasPermissions([
+    'permission1',
+    'permission4',
+  ]);
 
   t.is(hasPermissions, false);
 });
 
 test('info', async function(t) {
-
   const body = {
     id: 'id',
     make: 'make',
     model: 'model',
     year: 1234,
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/')
     .reply(200, body);
 
   const response = await vehicle.info();
   t.deepEqual(response, body);
-
 });
 
 test('location', async function(t) {
-
   const body = {
     latitude: 1234,
     longitude: 1234,
@@ -213,7 +208,8 @@ test('location', async function(t) {
   const headers = {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/location')
     .reply(200, body, headers);
 
@@ -222,28 +218,25 @@ test('location', async function(t) {
   t.true(_.isDate(response.age));
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
-
 });
 
 test('location - no age', async function(t) {
-
   const body = {
     latitude: 1234,
     longitude: 1234,
   };
   const headers = {};
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/location')
     .reply(200, body, headers);
 
   const response = await vehicle.location();
   t.deepEqual(response.data, body);
   t.is(response.age, null);
-
 });
 
 test('odometer', async function(t) {
-
   const body = {
     distance: 1234,
   };
@@ -251,7 +244,8 @@ test('odometer', async function(t) {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/odometer')
     .reply(200, body, headers);
 
@@ -261,18 +255,17 @@ test('odometer', async function(t) {
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('odometer - no age', async function(t) {
-
   const body = {
     distance: 1234,
   };
   const headers = {
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/odometer')
     .reply(200, body, headers);
 
@@ -280,11 +273,9 @@ test('odometer - no age', async function(t) {
   t.deepEqual(response.data, body);
   t.is(response.age, null);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('fuel', async function(t) {
-
   const body = {
     range: 1234,
     percentRemaining: 0.43,
@@ -294,7 +285,8 @@ test('fuel', async function(t) {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/fuel')
     .reply(200, body, headers);
 
@@ -304,11 +296,9 @@ test('fuel', async function(t) {
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('fuel - no age', async function(t) {
-
   const body = {
     range: 1234,
     percentRemaining: 0.43,
@@ -317,7 +307,8 @@ test('fuel - no age', async function(t) {
   const headers = {
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/fuel')
     .reply(200, body, headers);
 
@@ -325,18 +316,17 @@ test('fuel - no age', async function(t) {
   t.deepEqual(response.data, body);
   t.is(response.age, null);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('oil', async function(t) {
-
   const body = {
     lifeRemaining: 0.86,
   };
   const headers = {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/engine/oil')
     .reply(200, body, headers);
 
@@ -345,27 +335,24 @@ test('oil', async function(t) {
   t.true(_.isDate(response.age));
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
-
 });
 
 test('oil - no age', async function(t) {
-
   const body = {
     lifeRemaining: 0.86,
   };
   const headers = {};
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/engine/oil')
     .reply(200, body, headers);
 
   const response = await vehicle.oil();
   t.deepEqual(response.data, body);
   t.is(response.age, null);
-
 });
 
 test('tire pressure', async function(t) {
-
   const body = {
     frontLeft: 33.0,
     frontRight: 34.0,
@@ -376,7 +363,8 @@ test('tire pressure', async function(t) {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
     'sc-unit-system': 'imperial',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/tires/pressure')
     .reply(200, body, headers);
 
@@ -386,11 +374,9 @@ test('tire pressure', async function(t) {
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('tire pressure - no age', async function(t) {
-
   const body = {
     frontLeft: 33.0,
     frontRight: 34.0,
@@ -400,7 +386,8 @@ test('tire pressure - no age', async function(t) {
   const headers = {
     'sc-unit-system': 'imperial',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/tires/pressure')
     .reply(200, body, headers);
 
@@ -408,11 +395,9 @@ test('tire pressure - no age', async function(t) {
   t.deepEqual(response.data.tires, body);
   t.is(response.age, null);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('battery', async function(t) {
-
   const body = {
     range: 1234,
     percentRemaining: 0.43,
@@ -421,7 +406,8 @@ test('battery', async function(t) {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/battery')
     .reply(200, body, headers);
 
@@ -431,11 +417,9 @@ test('battery', async function(t) {
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('battery - no age', async function(t) {
-
   const body = {
     range: 1234,
     percentRemaining: 0.43,
@@ -443,7 +427,8 @@ test('battery - no age', async function(t) {
   const headers = {
     'sc-unit-system': 'metric',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/battery')
     .reply(200, body, headers);
 
@@ -451,11 +436,9 @@ test('battery - no age', async function(t) {
   t.deepEqual(response.data, body);
   t.is(response.age, null);
   t.is(response.unitSystem, headers['sc-unit-system']);
-
 });
 
 test('charge', async function(t) {
-
   const body = {
     isPluggedIn: true,
     state: 'CHARGING',
@@ -463,7 +446,8 @@ test('charge', async function(t) {
   const headers = {
     'sc-data-age': '2018-05-03T03:45:51+00:00',
   };
-  t.context.n = nocks.base()
+  t.context.n = nocks
+    .base()
     .get('/charge')
     .reply(200, body, headers);
 
@@ -472,11 +456,9 @@ test('charge', async function(t) {
   t.true(_.isDate(response.age));
   const expectedISOString = new Date(headers['sc-data-age']).toISOString();
   t.is(response.age.toISOString(), expectedISOString);
-
 });
 
 test('charge - no age', async function(t) {
-
   const body = {
     isPluggedIn: true,
     state: 'CHARGING',
@@ -527,82 +509,4 @@ test('unlock', async function(t) {
 
   t.deepEqual(_.xor(_.keys(response), ['status']), []);
   t.is(response.status, 'success');
-});
-
-test('batch request', async function(t) {
-  const defaultHeaders = {'sc-unit-system': 'imperial'};
-  const batchRequest = new BatchRequest(defaultHeaders);
-  batchRequest.addRequest('/odometer');
-  batchRequest.addRequest('/transmission/fluid', {'sc-unit-system': 'metric'});
-  batchRequest.addRequest('/fuel');
-  batchRequest.addRequest('/sunroof');
-  const requestBody = {
-    headers: {
-      'sc-unit-system': 'imperial',
-    },
-    requests: [
-      {
-        path: '/odometer',
-      },
-      {
-        headers: {
-          'sc-unit-system': 'metric',
-        },
-        path: '/transmission/fluid',
-      },
-      {
-        path: '/fuel',
-      },
-      {
-        path: '/sunroof',
-      },
-    ],
-  };
-  const responseBody = {
-    responses: [
-      {
-        headers: {'sc-unit-system': 'metric'},
-        path: '/odometer',
-        code: 200,
-        body: {
-          distance: 32768,
-        },
-      },
-      {
-        headers: {'sc-unit-system': 'imperial'},
-        path: '/transmission/fluid',
-        code: 200,
-        body: {
-          temperature: 98.2,
-          wear: 0.5,
-        },
-      },
-      {
-        headers: {'sc-unit-system': 'metric'},
-        path: '/fuel',
-        code: 200,
-        body: {
-          range: 550.8499755859375,
-          percentRemaining: 0.9449999928474426,
-        },
-      },
-      {
-        headers: {'sc-unit-system': 'metric'},
-        path: '/sunroof',
-        code: 501,
-        body: {
-          error: 'vehicle_not_capable_error',
-          message: 'Vehicle is not capable of performing request.',
-        },
-      },
-    ],
-  };
-
-  t.context.n = nocks
-    .base()
-    .post('/batch', requestBody)
-    .reply(200, responseBody);
-
-  const response = await vehicle.batch(batchRequest);
-  t.deepEqual(response.responses, responseBody.responses);
 });
